updated image generation code to account for change in AWS filenames in 2018
part of the filenames changed sometime in 2018 (changed from m3 to m6). Added an exception in the "get_s3_keys" function both in the notebook and in the py file so it properly loads the newer images. Also, increased the margin used to filter usable coordinates from the track data from 10 to 15. There were too many images with the horizon in them, and while there still are some with this extra margin threshold, there's still a healthy amount of detail visible.