made progress on tf.data Dataset
The dataset we need here is a combination of auxiliary data, image data, and multiple outputs, which is more complicated than any of the simple examples online. Formatting it correctly for use with a Tensorflow model is proving tricky, but this version is close to what we need. Still need to get the feature columns in there and build the model, but we're close now.