update "create_dataset" function
As far as I can tell, there are no existing examples online of anyone combining image data with numerical data using tf.data. So, I had to figure it out, which wasn't as simple as it sounds. It seems to be working here. Next update will include the creation of feature columns linked to this dataset, followed by the actual model itsself.