updated notebook 2 to use MaxPooling2D
updated notebook 2's model to use MaxPooling2D with Flatten instead of GlobalMaxPooling on MobileNet's output to see if performance improves. The GlobalMaxPooling layer was originally designed / tested for classification problems, so I figured this alternative approach might be worth it with a smaller kernel of (2,2). Turns out, I was right. Error is significantly reduced.