adding full disks, and updating notebooks
adding generated full disk images, and uploading new versions of the notebooks, with new, more sensible numbering. Also, the validation set I originally used was way too small due to an error in my sampling. Instead of sampling all records associated with N storms, I only pulled N samples, but only for the validation set. This is fixed now. This is why the validation error was so low before. The numbers are much more realistic now.